+++ Freezer/CHANGELOG.md
@@ -1,6 +1,13 @@
 ﻿# Changelog
 
-## v1.0.1 [2021-10-??]
+## v1.0.1 [2021-11-21]
+- Construction now requires some plastic
+- Added mod options
+  - Capactiy
+  - Energy saver power consumption
+  - Energy active power consumption
+
+## v1.0.1 [2021-10-10]
 - Fixed missing strings for automation port
 - Added Russian translation (done by |ScientisT|RU|)